Bash script to automate setup
# Please note that all these servers are no longer running and this was a task for a school project. 
Creates an MLN file that automates creating of OpenStack servers based on a couple of questions and installs and adds a couple of things.
Since this script was spammed it also had to check if servers with the same names were already running and shut them down if they were, with the nova command, and in foreman. 

When checks are done, it will build the mln project and spawn the servers. It will then clean puppet certs if already existing from previous servers, before it signs the new ones.